+++ headers/Config.hpp
@@ -1,0 +1,19 @@
+/**
+ * @title Hash Table
+ * @project hash_table
+ * @github https://github.com/chrismabon/hash_table
+ * @author Chris Mabon
+ */
+
+#ifndef HASH_TABLE_CONFIG_HPP
+#define HASH_TABLE_CONFIG_HPP
+
+/*
+ * TYPEDEFS
+ */
+
+ // integers
+ typedef unsigned short int USI_t;
+ typedef unsigned long long int ULLI_t;
+
+#endif //HASH_TABLE_CONFIG_HPP