+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Change Log
 
+## 2019-04-08
+
+#### v0.0.2
+
+###### Additions
+
+- tests for KeyNode class
+- graphic and missing README info
+- typedefs for large and small unsigned integers
+- progress indicator chart for various features
+
 ## 2019-03-21
+
+#### v0.0.2
+
+###### Additions
+
+- Add print node and print list functions
 
 #### v0.0.1
 
@@ -8,7 +25,6 @@
 
 - Add comments and documentation to node, list,
 table,  output classes
-
 - Add TODOs marking where and what functions to implement
 
 ## 2019-03-20
@@ -19,7 +35,6 @@ table,  output classes
 
 - Add classes, class members, and associated gets/sets,
 constructor, and destructor
-
 - Add doc and testing files
 
 #### Initial Commit